requirements.txt
pandas
PyMuPDF
pytesseract
pdf2image
python-dotenv
neo4j
loguru
pydantic
openai
langchain
langgraph
langchain-openai
langchain-ollama
langfuse